Classes divorced from their abstractions
Classes behave as expected as they used to with the same structure as they did, but now they aren't tied to any of their superclasses, as they no longer exist in the project. This is a good starting point to try to refactor the eyes to work in accordance with the new (simpler) architecture.